feature/#158/threaded-syscalls: Seperated user space code from kernel space code.

+++ src/system/include/kernel/process.h
@@ -61,7 +61,7 @@ extern "C" {
 #define _IMPORT_PROGRAM(_name_)\
         _PROGMAN_EXTERN_C const size_t __builtin_app_##_name_##_gs__;\
         _PROGMAN_EXTERN_C const size_t __builtin_app_##_name_##_ss__;\
-        _PROGMAN_EXTERN_C size_t __builtin_app_##_name_##_main(int, char**)
+        _PROGMAN_EXTERN_C int __builtin_app_##_name_##_main(int, char**)
 
 #define int_main(_name_, stack_depth, argc, argv)\
         _PROGMAN_CXX const size_t __builtin_app_##_name_##_gs__ = sizeof(struct _GVAR_STRUCT_NAME);\
